Refactored & moved things around
