+++ src/components/server/ElementViewer.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import TST from '@/components/server/transitions/TextSlideTransitionElementViewer.vue'
 import { currentMapRawType } from '@/ts/manager/CurrentMap';
-import { ElementViewerComponent } from '@/ts/servers/CurrentServerRedirect';
+import { ElementViewerComponent } from '@/ts/server/CurrentServerRedirect';
 </script>
