--- src/ts/data/IServerSearch.ts
@@ -1,9 +1,0 @@
-import { ServerMap } from "./ServerMap";
-import { TagNode } from "./Tag";
-
-export interface IServerSearch<T extends ServerMap> {
-    validTags: string[];
-
-    grabTags(tagNode: TagNode): void;
-    isMapGood(map: T): boolean;
-}
