+++ src/ts/servers/mineplex/MineplexData.ts
@@ -1,5 +1,5 @@
 import { MineplexMap as M } from "./MineplexMap";
-import { Download as D } from "@/ts/data/ServerMap";
+import { Download as D } from "@/ts/server/AServerMap";
 
 const Z = "Main download";
 const K: string[] = [
