+++ src/views/Server.vue
@@ -2,7 +2,7 @@
 import ElementViewer from '@/components/server/ElementViewer.vue'
 import SearchBar from '@/components/server/searchbar/SearchBar.vue'
 import SearchGrid from '@/components/server/grid/Grid.vue'
-import { parseUrlArgs } from '@/ts/utils/ArgsHandler.js';
+import { parseUrlArgs } from '@/ts/manager/UrlArgsHandler';
 
 parseUrlArgs()
